Cleaned the create and delete role tests and added helper functions
